+++ authority-portal-frontend/src/app/core/services/config/profiles/ui-theme-config.ts
@@ -23,5 +23,11 @@ export type UiThemeConfig = Pick<
   | 'brandLogoSrc'
   | 'brandLogoSmallSrc'
   | 'brandFaviconSrc'
+  | 'connectorSelfOwnedIconSrc'
+  | 'connectorSelfOwnedIconStyle'
+  | 'connectorCaasIconSrc'
+  | 'connectorCaasIconStyle'
+  | 'caasResellerBrandLogoSrc'
+  | 'caasResellerBrandLogoStyle'
   | 'copyrightCompanyName'
 >;